+++ aws_controller/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class AwsControllerConfig(AppConfig):
+    name = 'aws_controller'
