+++ vision_controller/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class VisionControllerConfig(AppConfig):
+    name = 'vision_controller'
